[BUGFIX] Rewrite namespace paths based on distUrl
Usage in real live instances revealed broken
autoload information based on the fact that
the empty source url has been used instead
of the distUrl.

This change hardens the processing and used
now the correct value to determine package
paths. Tests based on wrong assumptions are
fixed along the way.